Run seqkit

Run seqkit

Usage

seqkit(subcmd = NULL, ..., seqkit = NULL)

Arguments

subcmd

Sub-Command of seqkit.

...

<dynamic dots> Additional arguments passed to seqkit subcmd command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using shQuote(). Details see: cmd_help(seqkit subcmd()).

seqkit

A string of path to seqkit command.

Value

A command object.

See also