TRUST4: immune repertoire reconstruction from bulk and single-cell RNA-seq data

TRUST4: immune repertoire reconstruction from bulk and single-cell RNA-seq data

Usage

trust4(
  file1,
  ref_coordinate,
  ...,
  file2 = NULL,
  mode = NULL,
  ref_annot = NULL,
  ofile = NULL,
  odir = getwd(),
  trust4 = NULL
)

trust4_imgt_annot(
  species = "Homo_sapien",
  ...,
  ofile = "IMGT+C.fa",
  odir = getwd(),
  perl = NULL
)

trust4_gene_names(imgt_annot, ofile = "bcr_tcr_gene_name.txt", odir = getwd())

Arguments

file1

Path to bam file or fastq file.

ref_coordinate

Path to the fasta file coordinate and sequence of V/D/J/C genes.

...

trust4: <dynamic dots> Additional arguments passed to run-trust4 command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using shQuote(). Details see: cmd_help(run-trust4()).

trust4_imgt_annot: <dynamic dots> Additional arguments passed to trust4_imgt_annot command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using shQuote(). Details see: cmd_help(trust4_imgt_annot()).

file2

Path to the second paired-end read fastq file, only used for mode = "fastq".

mode

One of "bam" or "fastq". If NULL, will be inferred from file1.

ref_annot

Path to detailed V/D/J/C gene reference file, such as from IMGT database. (default: not used). (recommended).

ofile

trust4: Prefix of output files. (default: inferred from file prefix).

trust4_imgt_annot: Output file name.

trust4_gene_names: Output file name.

odir

A string of path to the output directory.

trust4

A string of path to run-trust4 command.

species

Species to extract IMGT annotation, details see https://www.imgt.org//download/V-QUEST/IMGT_V-QUEST_reference_directory/.

perl

A string of path to perl command.

imgt_annot

Path of IMGT annotation file, created via trust4_imgt_annot.

Value

A command object.

See also